Internal

This chapter documents internal functions that are typically only for use of the developers of package Modia.

Code Generation

This section provides functions to generate Julia code of the transformed equations.

Modia.SimulationModel — Type

simulationModel = SimulationModel{FloatType,TimeType}(
        modelModule, modelName, getDerivatives!, equationInfo, x_startValues,
        parameters, timeName, w_invariant_names;
        vSolvedWithInitValuesAndUnit::OrderedDict{String,Any}(),
        vEliminated::Vector{Int}=Int[],
        vProperty::Vector{Int}=Int[],
        var_name::Function = v->nothing)

Arguments

modelModule: Module in which @instantiateModel is invoked (it is used for Core.eval(modelModule, ...)), that is evaluation of expressions in the environment of the user.

modelName::String: Name of the model

getDerivatives::Function: Function that is used to evaluate the model equations, typically generated with [Modia.generate_getDerivatives!].

equationInfo::Modia.EquationInfo: Information about the states and the equations.

x_startValues:: Deprecated (is no longer used).

parameters: A hierarchical NamedTuple of (key, value) pairs defining the parameter and init/start values.

timeName: Name of time (as Symbol)

w_invariant_names: A vector of variable names (as vector of symbols or Expr)

Modia.generate_getDerivatives! — Function

code = generate_getDerivatives!(AST, equationInfo, parameters, timeName, w_invariant_names, functionName;
                                hasUnits=false)

Return the code of the getDerivatives! function as Expr using the Symbol functionName as function name. By eval(code) or fc = @RuntimeGeneratedFunction(code) the function is compiled and can afterwards be called.

Arguments

AST::Vector{Expr}: Abstract Syntax Tree of the equations as vector of Expr.

equationInfo::Modia.EquationInfo: Data structure returned by `Modia.getSortedAndSolvedAST holding information about the states.

parameters: Vector of parameter names (as vector of symbols)

timeName: Name of time (as symbol)

w_invariant_names: Vector of variable names (as vector of symbols or Expr).

functionName::Function: The name of the function that shall be generated.

Optional Arguments

pre:Vector{Symbol}: pre-variable names

hasUnits::Bool: = true, if variables have units. Note, the units of the state vector are defined in equationinfo.

Modia.init! — Function

success = init!(simulationModel)

Initialize simulationModel::SimulationModel at startTime. In particular:

Empty result data structure.

Merge parameter and init/start values into simulationModel.

Construct x_start.

Call simulationModel.getDerivatives! once with isInitial(simulationModel) = true to compute and store all variables in the result data structure at startTime and initialize simulationModel.linearEquations.

Check whether explicitly solved variables that have init-values defined, have the required value after initialization (-> otherwise error).

If initialization is successful return true, otherwise false.

Modia.outputs! — Function

outputs!(x, t, integrator)

DifferentialEquations FunctionCallingCallback function for SimulationModel that is used to store results at communication points.

Modia.terminate! — Function

terminate!(m::SimulationModel, x, time)

Terminate model.

Modia.derivatives! — Function

derivatives!(derx, x, m, t)

DifferentialEquations callback function to get the derivatives.

Modia.DAEresidualsForODE! — Function

DAEresidualsForODE!(residuals, derx, x, m, t)

DifferentialEquations callback function for DAE integrator for ODE model

Modia.affectEvent! — Function

affectEvent!(integrator, stateEvent, eventIndex)

Called when a time event (stateEvent=false) or state event (stateEvent=true) is triggered. In case of stateEvent, eventIndex is the index of the crossing function that triggered the event.

Modia.zeroCrossings! — Function

zeroCrossings!(z, x, t, integrator)

Called by integrator to compute zero crossings

Modia.affectStateEvent! — Function

affectStateEvent!(integrator, event_index)

Called by integrator when a state event is triggered

Modia.timeEventCondition! — Function

timeEventCondition!(u, t, integrator)

Called by integrator to check if a time event occurred

Modia.affectTimeEvent! — Function

affectTimeEvent!(integrator)

Called by integrator when a time event is triggered

Modia.addToResult! — Function

addToResult!(instantiatedModel, x, time, w_invariant...)

Add result of current time instant (time, x, der_x, w_invariant, w_segmented) to instantiatedModel.

Modia.getFloatType — Function

floatType = getFloatType(simulationModel::SimulationModel)

Return the floating point type with which simulationModel is parameterized (for example returns: Float64, Float32, DoubleFloat, Measurements.Measurement{Float64}).

Modia.measurementToString — Function

str = measurementToString(v)

Return variable v::Measurements.Measurement{FloatType} or a vector of such variables in form of a string will the full number of significant digits.

Inquiries in Model

The functions in this section can be called in the model code or in functions that are called from the model code.

Modia.isInitial — Function

isInitial(instantiatedModel)

Return true, if initialization phase of simulation.

Modia.isTerminal — Function

isTerminal(instantiatedModel)

Return true, if terminal phase of simulation.

Modia.isEvent — Function

isEvent(instantiatedModel)

Return true, if event phase of simulation (including initialization).

Modia.isFirstEventIteration — Function

isFirstEventIteration(instantiatedModel)

Return true, if event phase of simulation (including initialization) and during the first iteration of the event iteration.

Modia.isFirstEventIterationDirectlyAfterInitial — Function

isFirstEventIterationDirectlyAfterInitial(instantiatedModel)

Return true, if first iteration directly after initialization where initial=true (so at the startTime of the simulation).

Modia.isAfterSimulationStart — Function

isAfterSimulationStart(instantiatedModel)

Return true, if after start of simulation (returns false during initialization).

Modia.isZeroCrossing — Function

isZeroCrossing(instantiatedModel)

Return true, if event indicators (zero crossings) shall be computed.

Modia.storeResults — Function

storeResults(instantiatedModel)

Return true, if results shall be stored.

Modia.getTime — Function

tCurrent = getTime(instantiatedModel)

Return current simulation time.

Variable definitions in functions

The following functions can be used to define states and algebraic variables inside functions:

Modia.new_x_segmented_variable! — Function

index = new_x_segmented_variable!(
            instantiatedModel::SimulationModel,
            x_name::String, der_x_name::String, startOrInit, x_unit::String="";
            nominal::Float64 = NaN, unbounded::Bool = false)::Int

Reserves storage location for a new xsegmented and derxsegmented variable and returns the index (= xsegmented_startIndex) to access this storage location, in particular

to copy state values from instantiatedModel.x_segmented[index:index+prod(dims(startOrInit))-1] into this storage location

to copy state derivative values of this storage location to instantiatedModel.derxsegmented[index:index+prod(dims(startOrInit))-1]

Value startOrInit is the start/init value used during re-initialization of the new segment with initFullRestart!(..).

Modia.new_w_segmented_variable! — Function

index = new_w_segmented_variable!(partiallyInstantiatedModel::SimulationModel, name::String,
                                  w_segmented_default, unit::String="")::Int

Reserve storage location for a new wsegmented variable. The returned index is used to store the wsegmented value at communication points in the result data structure. Value wsegmenteddefault is stored as default value and defines type and (fixed) size of the variable in this segment.

Modia.new_alias_segmented_variable! — Function

new_alias_segmented_variable!(partiallyInstantiatedModel, name, aliasName, aliasNegate=false)

Define new alias segmented variable.

Modia.new_z_segmented_variable! — Function

startIndex = new_z_segmented_variable!(instantiatedModel, nz)

Reserve storage location for nz new segmented zero crossing function and return the startIndex to copy it in the vectors of zero crossings

Modia.get_x_startIndex_from_x_segmented_startIndex — Function

x_startIndex = get_x_startIndex_from_x_segmented_startIndex(instantiatedModel::SimulationModel, x_segmented_startIndex)

Return the startindex of an xsegmented state with respect to the x-vector, given the startIndex with respect to the xsegmented vector (xsegmentedstartIndex is the return value of newxsegmented_variable!(..)).

Modia.get_scalar_x_segmented_value — Function

value = Modia.get_scalar_x_segmented_value(instantiatedModel, startIndex)

Return scalar segmented state value from instantiatedModel given startIndex (returned from new_x_segmented_variable!(..)).

Modia.get_SVector3_x_segmented_value — Function

value = Modia.get_SVector3_x_segmented_value(instantiatedModel, startIndex)

Return SVector{3,FloatType}(..) segmented state value from instantiatedModel given startIndex (returned from new_x_segmented_variable!(..)).

Modia.get_Vector_x_segmented_value! — Function

Modia.get_Vector_x_segmented_value!(instantiatedModel::SimulationModel, startIndex, xi::Vector{FloatType})::Nothing

Copy state from instantiatedModel at index startIndex into pre-allocated vector xi.

Modia.add_der_x_segmented_value! — Function

Modia.add_der_x_segmented_value!(instantiatedModel, startIndex, der_x_segmented_value::[FloatType|Vector{FloatType}])

Copy scalar or array segmented state derivative value der_x_segmented_value into instantiatedModel starting at index startIndex (returned from new_x_segmented_variable!(..)).

Modia.add_w_segmented_value! — Function

Modia.add_w_segmented_value!(instantiatedModel::SimulationModel, index::Int, w_segmented_value)::Nothing

Store deepcopy(wsegmentedvalue) at index in instantiatedModel.